+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Adds `ExtraSmall` and `ExtraExtraSmall` size variants to the `Image` component.
 - Adds `ExtraSmall` and `ExtraExtraSmall` size variants for `CardImage` in the `Card` component.
 - Adds `AlertNotificationImportant` icon to `Icon` component.
+- Adds `Toggle` component
 
 ### Changes